accept option as args, consolidate tests, more efficient glob match
