+++ README.md
@@ -34,7 +34,7 @@ module.exports = {
 That will do it! By default it will match any file with a `jade` extension, and use the pushState override on it. However, if you want it to match a different pattern, like maybe only jade files in one specific folder, you can just pass a string with a glob matcher as such:
 
 ```javascript
-new PushState('**/templates/*.jade')
+new PushState({ files: '**/templates/*.jade' })
 ```
 
 ## License & Contributing
