+++ lib/index.js
@@ -12,8 +12,8 @@ module.exports = class PushState {
    * @param {String} opts.root - project root
    * @param {Array} opts.dumpDirs - directories to dump to public
    */
-  constructor (opts) {
-    this.files = opts || '**/*.jade'
+  constructor (opts = {}) {
+    this.files = opts.files || '**/*.jade'
   }
 
   apply (compiler) {
@@ -25,9 +25,7 @@ module.exports = class PushState {
       const jadeFiles = compiler.options.spike.files.jade
 
       // filter in files matching the user-defined matcher
-      const matchingFiles = jadeFiles.filter((f) => {
-        return mm.isMatch(f.replace(`${util.conf.context}/`, ''), this.files)
-      })
+      const matchingFiles = util.matchGlobs(jadeFiles, this.files)
 
       // now we are going to build the link injector script
       let mod = ''
@@ -35,7 +33,7 @@ module.exports = class PushState {
       // first, we require in all the matching jade files so that we have the
       // contents ready to inject if needed
       matchingFiles.forEach((f) => {
-        mod += `exports['${util.getOutputPath(f).replace(/\.jade/, '')}'] = require('../${f.replace(`${util.conf.context}/`, '')}')\n`
+        mod += `exports['${util.getOutputPath(f).relative.replace(/\.jade/, '')}'] = require('../${f.replace(`${util.conf.context}/`, '')}')\n`
       })
 
       // now we add the script that matches link clicks to templates
@@ -52,7 +50,7 @@ module.exports = class PushState {
 
       // finally, we add that file as an entry to webpack so that it shows up
       util.addFilesAsWebpackEntries(compilation, [tempFile])
-        .done(() => done())
+        .done(() => done(), done)
     })
 
     compiler.plugin('compilation', (compilation) => {
