+++ test/index.js
@@ -24,3 +24,24 @@ test.cb('compiles correctly', (t) => {
 
   project.compile()
 })
+
+test.cb('filters files according to files option', (t) => {
+  const project = new Spike({
+    root: path.join(fixtures, 'string'),
+    entry: { main: ['./main.js'] },
+    plugins: [new PushState({ files: '**/foo.jade' })]
+  })
+
+  project.on('error', t.end)
+  project.on('warning', t.end)
+  project.on('compile', () => {
+    const build = fs.readFileSync(path.join(fixtures, 'string/public/main.js'), 'utf8')
+    t.truthy(build.match(/<p>bar<\/p>/))
+    t.truthy(build.match(/exports\['foo'\] =/))
+    t.falsy(build.match(/exports\['index'\] =/))
+    t.falsy(build.match(/<p>index<\/p>/))
+    t.end()
+  })
+
+  project.compile()
+})
